[client] refactor: split app_panels.rs into one file per panel
  Replaces the 1050-line monolithic app_panels.rs with a panels/ module
  containing one file per panel: bookmarks, operator, radio_control,
  radios, server, source_control, waterfall_control. panels/mod.rs holds
  only draw_left_panel. No behaviour change.
