+++ rigflow-client/src/ui/mod.rs
@@ -13,4 +13,4 @@ pub mod waterfall;
 mod app_actions;
 mod app_center;
 mod app_dialogs;
-mod app_panels;
+mod panels;
